tegra: usb: prevent crash if irq occurs while sleeping
In some rare cases, EHCI irqs interrupts be triggered while the block is
suspended. This is notably the case when unplugging a host cable to
which no device is attached. In this case, the interrupt handler will be
called with the block powered off, which will eventually crash. This
patch workarounds this by returning from the interrupt handler
immediatly if the block is not powered on.

Change-Id: I7763f56f20a984e3db716306b6bf8ec223c90363
Signed-off-by: Alexandre Courbot <acourbot@nvidia.com>
Reviewed-on: http://git-master/r/35604
Reviewed-by: Suresh Mangipudi <smangipudi@nvidia.com>
Reviewed-by: Hanumanth Venkateswa Moganty <vmoganty@nvidia.com>