video: tegra: dc: update emc clk for multiple dcs
When an overlay is opened or closed it updates emc clk as per pclk.
However, this does not consider state of other dc for memory bandwidth
consideration. This results in underflow errors when both display
controllers are in use. To avoid underflow, update emc clk as per
state of both dcs.

BUG 865578

Change-Id: I595c17a2284ce9857acb040e3c6ee6cadf0b4e50
Reviewed-on: http://git-master/r/50736
Tested-by: Nitin Kumbhar <nkumbhar@nvidia.com>
Reviewed-by: Bharat Nihalani <bnihalani@nvidia.com>
Tested-by: Gerrit_Virtual_Submit
Reviewed-by: Jon Mayo <jmayo@nvidia.com>